+++ hazelcast/src/main/java/com/hazelcast/map/impl/JsonMetadataInitializer.java
@@ -19,7 +19,7 @@ package com.hazelcast.map.impl;
 import com.fasterxml.jackson.core.JsonFactory;
 import com.fasterxml.jackson.core.JsonParser;
 import com.hazelcast.core.HazelcastJsonValue;
-import com.hazelcast.json.internal.JsonSchemaHelper;
+import com.hazelcast.internal.json.query.JsonSchemaHelper;
 import com.hazelcast.nio.serialization.Data;
 
 import java.io.ByteArrayInputStream;
