+++ hazelcast/src/test/java/com/hazelcast/internal/json/query/JsonMetadataCreationMigrationTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.json.internal;
+package com.hazelcast.internal.json.query;
 
 import com.hazelcast.config.Config;
 import com.hazelcast.config.MetadataPolicy;
