+++ hazelcast/src/test/java/com/hazelcast/internal/json/query/JsonSchemaHelperNullTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.json.internal;
+package com.hazelcast.internal.json.query;
 
 import com.hazelcast.config.InMemoryFormat;
 import com.hazelcast.core.HazelcastJsonValue;
