+++ hazelcast/src/test/java/com/hazelcast/query/impl/getters/AbstractJsonGetterTest.java
@@ -20,8 +20,8 @@ import com.fasterxml.jackson.core.JsonFactory;
 import com.hazelcast.core.HazelcastJsonValue;
 import com.hazelcast.internal.json.Json;
 import com.hazelcast.internal.json.JsonObject;
-import com.hazelcast.json.internal.JsonSchemaHelper;
-import com.hazelcast.json.internal.JsonSchemaNode;
+import com.hazelcast.internal.json.query.JsonSchemaHelper;
+import com.hazelcast.internal.json.query.JsonSchemaNode;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.annotation.ParallelTest;
 import com.hazelcast.test.annotation.QuickTest;
